#14 refactor
